Debug error in sampling trajectories
Mistakenly using wrong input arguments. To be clearer, here are some terminologies to remember.

batch size: can refer to 'the number of transitions' or 'the number of trajectories/episodes'. In this code, the batch size refers to the number of transitions we collect, i.e., (the number of trajectories/episodes) x (the length of each trajectory/episode) - or the total sum of lengths of all trajectories/episodes.

Previously, the batch size / episode length was used. This determines the number of trajectories, not the min_timesteps_per_batch for smple_trajectories. Either using sample_trajectories(..., batch_size, ...) or sample_n_trajectories (..., batch size / episode length, ...)