Debug sampling methods in rl_trainer.py
Wrong use of sample_trajectories was fixed. Hence, additional check made to the replay_buffer.py is unnecessary and removed.

Note: batch_size can be either 'num of transitions' or 'num of trajectories/episodes'. The 'n of transitions' is basically the total timestep, i.e., sum of length of all episodes.